+++ arangod/Aql/Optimizer/Rule/DistributeInCluster.h
@@ -24,10 +24,25 @@
 
 #pragma once
 
+#include "Aql/ExecutionNode/DistributeNode.h"
+#include "Aql/ExecutionNode/GatherNode.h"
 #include "Aql/ExecutionPlan.h"
 
 namespace arangodb::aql {
 class Optimizer;
+class SubqueryNode;
+
+//// @brief create a DistributeNode for the given ExecutionNode
+DistributeNode* createDistributeNodeFor(ExecutionPlan& plan,
+                                        ExecutionNode* node);
+
+//// @brief create a gather node matching the given DistributeNode
+GatherNode* createGatherNodeFor(ExecutionPlan& plan, DistributeNode* node);
+
+//// @brief enclose a node in DISTRIBUTE/GATHER
+DistributeNode* insertDistributeGatherSnippet(ExecutionPlan& plan,
+                                              ExecutionNode* at,
+                                              SubqueryNode* snode);
 
 /// @brief distribute operations in cluster
 void distributeInClusterRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
